refactor: 重构 TimePicker 脚本
